third_party/blink/web_tests/fast/dom/shadow/select-in-shadowdom-expected.txt

CONSOLE WARNING: line 18: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
When selecting from non-anchor Node to anchor node in ShadowDOM, client should not cause page jump.

Selecting from a node to another node in ShadowDOM. This should not start page navigation.
PASS lastClickTarget is host


Clicking a node in ShadowDOM.
PASS lastClickTarget is host


PASS successfullyParsed is true

TEST COMPLETE
